+++ packages/dendron-next-server/hooks/useApplyGraphConfig.tsx
@@ -1,7 +1,9 @@
 import cytoscape from "cytoscape";
 import { useEffect } from "react";
+import { createLogger } from "@dendronhq/common-frontend";
 import { getEulerConfig } from "../components/graph";
 import { GraphConfig, GraphElements } from "../lib/graph";
+import _ from "lodash";
 
 const useApplyGraphConfig = ({
   graph,
@@ -12,6 +14,7 @@ const useApplyGraphConfig = ({
   config: GraphConfig;
   elements: GraphElements;
 }) => {
+  const logger = createLogger("Graph: useApplyGraphConfig");
   const { nodes, edges } = elements;
   const isLargeGraph = nodes.length + Object.values(edges).flat().length > 1000;
 
@@ -55,18 +58,102 @@ const useApplyGraphConfig = ({
         const vaultClass = `vault-${vaultName}`;
 
         const includedElements = graph.$(`.${vaultClass}`);
-        const elementCount = includedElements.length;
 
         // If elements should be included
-        if (v?.value && includedElements.hasClass("hidden")) {
-          includedElements.removeClass("hidden");
+        if (v?.value && includedElements.hasClass("hidden--vault")) {
+          includedElements.removeClass("hidden--vault");
         }
 
         // If elements should not be included
-        else if (!v?.value && !includedElements.hasClass("hidden")) {
-          includedElements.addClass("hidden");
+        else if (!v?.value && !includedElements.hasClass("hidden--vault")) {
+          includedElements.addClass("hidden--vault");
         }
       });
+  };
+  const applyFilterRegexConfig = () => {
+    if (!graph || graph.$("*").length === 0) return;
+
+    const regexTypes: ("whitelist" | "blacklist")[] = [
+      "whitelist",
+      "blacklist",
+    ];
+
+    // Process the whitelist and blacklist inputs
+    regexTypes.forEach((type) => {
+      const classNameHidden = `hidden--regex-${type}`;
+      const regexItem =
+        config[
+          `filter.regex-${type}` as
+            | "filter.regex-whitelist"
+            | "filter.regex-blacklist"
+        ];
+
+      // Accept comma-separated or space-separated lists
+      const regexItemInputs = regexItem.value.split(/(,| )/);
+
+      // Form cytoscape selectors from input
+      const matchingInput = regexItemInputs.reduce((acc, input, i) => {
+        const prefix = i === 0 ? "" : ", ";
+        const trimmedInput = input.trim();
+        if (trimmedInput === "") return acc;
+
+        return (acc += `${prefix}[fname *= "${trimmedInput}"], [vault *= "${trimmedInput}"], [label *= "${trimmedInput}"]`);
+      }, "");
+
+      const excludedInput = regexItemInputs.reduce((acc, input, i) => {
+        const trimmedInput = input.trim();
+        if (trimmedInput === "") return acc;
+
+        return (acc += `[fname !*= "${trimmedInput}"][vault !*= "${trimmedInput}"][label !*= "${trimmedInput}"]`);
+      }, "");
+
+      const matchingElements = graph.$(matchingInput);
+      const excludedElements = graph.$(excludedInput);
+
+      const hideElement = (element: cytoscape.SingularElementReturnValue) => {
+        if (!element.hasClass(classNameHidden)) {
+          element.addClass(classNameHidden);
+        }
+      };
+      const showElement = (element: cytoscape.SingularElementReturnValue) => {
+        if (element.hasClass(classNameHidden)) {
+          element.removeClass(classNameHidden);
+        }
+      };
+
+      matchingElements.forEach((element) => {
+        if (type === "whitelist") showElement(element);
+        if (type === "blacklist") hideElement(element);
+      });
+
+      // If no input,
+      if (
+        regexItem.value === "" ||
+        (type === "whitelist" && excludedElements.length === 0) ||
+        (type === "blacklist" && matchingElements.length === 0)
+      ) {
+        graph.$("*").removeClass(classNameHidden);
+        return;
+      }
+
+      excludedElements.forEach((element) => {
+        if (type === "whitelist") hideElement(element);
+        if (type === "blacklist") showElement(element);
+      });
+    });
+  };
+  const applyFilterStubsConfig = () => {
+    if (!graph || graph.$("*").length === 0) return;
+    if (_.isUndefined(config["filter.show-stubs"])) return;
+
+    const configItem = config["filter.show-stubs"];
+
+    // If should show stubs
+    if (configItem.value) {
+      graph.$("[?stub]").removeClass("hidden--stub");
+    } else {
+      graph.$("[?stub]").addClass("hidden--stub");
+    }
   };
 
   const applyConfig = () => {
@@ -74,6 +161,10 @@ const useApplyGraphConfig = ({
 
     applyDisplayConfig();
     applyVaultConfig();
+    applyFilterRegexConfig();
+    applyFilterStubsConfig();
+
+    logger.log(graph.$(".hidden--regex"));
 
     graph.layout(getEulerConfig(!isLargeGraph)).run();
   };